Make quicktype callable on the web
