+++ package.json
@@ -2,8 +2,8 @@
   "name": "quicktype",
   "version": "4.1.1",
   "license": "Apache-2.0",
-  "main": "quicktype.js",
-  "types": "quicktype.d.ts",
+  "main": "index.js",
+  "types": "index.d.ts",
   "repository": "https://github.com/quicktype/quicktype",
   "scripts": {
     "pub": "script/publish.sh",
@@ -23,6 +23,7 @@
     "pluralize": "^7.0.0",
     "stream-json": "0.5.2",
     "string-hash": "^1.1.3",
+    "string-to-stream": "^1.1.0",
     "unicode-properties": "quicktype/unicode-properties#dist"
   },
   "devDependencies": {
